customer segmentation/requirements.txt
pandas
kneed
numpy
scikit-learn
matplotlib
tensorflow
streamlit
flask
twilio
python-dotenv
mlxtend

flask==2.0.1
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
lightgbm==3.2.1
python-dotenv==0.19.0
twilio==7.0.0
openpyxl==3.0.7